Refactored lib/auth/services/auth_provider.dart -> lib/auth/services/auth_services/auth_provider.dart
