+++ lib/auth/bloc/auth_bloc.dart
@@ -1,4 +1,4 @@
-import 'package:echno_attendance/auth/services/auth_provider.dart';
+import 'package:echno_attendance/auth/services/auth_services/auth_provider.dart';
 import 'package:echno_attendance/auth/bloc/auth_event.dart';
 import 'package:echno_attendance/auth/bloc/auth_state.dart';
 import 'package:bloc/bloc.dart';
