+++ lib/auth/services/auth_service.dart
@@ -1,4 +1,4 @@
-import 'package:echno_attendance/auth/services/auth_provider.dart';
+import 'package:echno_attendance/auth/services/auth_services/auth_provider.dart';
 import 'package:echno_attendance/auth/services/auth_user.dart';
 import 'package:echno_attendance/auth/services/firebase_auth_provider.dart';
