+++ test/auth_test.dart
@@ -1,5 +1,5 @@
 import 'package:echno_attendance/auth/utilities/auth_exceptions.dart';
-import 'package:echno_attendance/auth/services/auth_provider.dart';
+import 'package:echno_attendance/auth/services/auth_services/auth_provider.dart';
 import 'package:echno_attendance/auth/services/auth_user.dart';
 import 'package:test/test.dart';
